+++ src/test/java/de/gematik/pki/error/ErrorClassifierTest.java
@@ -14,28 +14,16 @@
  * limitations under the License.
  */
 
-package de.gematik.pki.utils;
+package de.gematik.pki.error;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
-import static org.junit.jupiter.api.Assertions.assertNull;
-import de.gematik.pki.exception.GemPkiException;
-import de.gematik.pki.ocsp.OcspConstants;
-import java.nio.file.Path;
 import org.junit.jupiter.api.Test;
 
-class P12ReaderTest {
+class ErrorClassifierTest {
 
     @Test
-    void verifyGetContentFromP12Valid() {
-        assertDoesNotThrow(() ->
-            P12Reader.getContentFromP12(OcspConstants.P12_OCSP_RESPONSE_SIGNER_RSA, OcspConstants.P12_PASSWORD)
-        );
+    void getDescription() {
+        assertDoesNotThrow(ErrorClassifier.SECURITY_ERROR::getDescription);
     }
 
-    @Test
-    void verifyGetContentFromP12Null() throws GemPkiException {
-        assertNull(
-            P12Reader.getContentFromP12(Path.of("src/test/resources/certificates/empty.p12"), OcspConstants.P12_PASSWORD)
-        );
-    }
 }