+++ src/test/java/de/gematik/pki/ocsp/OcspVerifierTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 gematik GmbH
+ * Copyright (c) 2022 gematik GmbH
  * 
  * Licensed under the Apache License, Version 2.0 (the License);
  * you may not use this file except in compliance with the License.
@@ -16,6 +16,7 @@
 
 package de.gematik.pki.ocsp;
 
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import de.gematik.pki.exception.GemPkiException;
 import de.gematik.pki.utils.CertificateProvider;
@@ -52,6 +53,13 @@ class OcspVerifierTest {
             .gen(ocspReq);
 
         assertThat(OcspVerifier.isStatusGood(ocspResp)).isTrue();
+    }
+
+    @Test
+    void nonNullTests() {
+        assertThatThrownBy(
+            () -> OcspVerifier.isStatusGood(null))
+            .isInstanceOf(NullPointerException.class);
     }
 
 }