+++ src/test/java/de/gematik/pki/tsl/TspInformationProviderTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 gematik GmbH
+ * Copyright (c) 2022 gematik GmbH
  * 
  * Licensed under the Apache License, Version 2.0 (the License);
  * you may not use this file except in compliance with the License.
@@ -25,7 +25,6 @@ import de.gematik.pki.utils.CertificateProvider;
 import de.gematik.pki.utils.ResourceReader;
 import eu.europa.esig.trustedlist.jaxb.tsl.TrustStatusListType;
 import java.io.IOException;
-import java.net.URISyntaxException;
 import java.security.cert.X509Certificate;
 import java.util.Optional;
 import lombok.SneakyThrows;
@@ -101,7 +100,7 @@ class TspInformationProviderTest {
     }
 
     @Test
-    void generateTspServiceSubsetServiceSupplyPointMissing() throws GemPkiException, URISyntaxException, IOException {
+    void generateTspServiceSubsetServiceSupplyPointMissing() throws GemPkiException, IOException {
         final Optional<TrustStatusListType> tslAltCaMissingSsp = TslReader
             .getTsl(ResourceReader.getFilePathFromResources("tsls/defect/TSL_defect_altCA_missingSsp.xml"));